LUA: update documentation of readbyterange to reflect what it actually does. Add readbyterangetable which allows the user to select the starting index to any desired value. Resolves #2844

+++ src/BizHawk.Client.Common/lua/NLuaTableHelper.cs
@@ -39,7 +39,7 @@ namespace BizHawk.Client.Common
 
 		public IEnumerable<T> EnumerateValues<T>(LuaTable table) => table.Values.Cast<T>();
 
-		public LuaTable ListToTable<T>(IList<T> list, int indexFrom = 0)
+		public LuaTable ListToTable<T>(IList<T> list, long indexFrom = 0)
 		{
 			var table = _lua.NewTable();
 			for (int i = 0, l = list.Count; i != l; i++) table[indexFrom + i] = list[i];
